+++ app/src/main/res/layout/activity_more_all_notice.xml
@@ -34,7 +34,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/tv_notice_all_all_notice"
-            tools:listitem="@layout/list_notice" />
+            tools:listitem="@layout/list_all_notice" />
 
 
     </androidx.constraintlayout.widget.ConstraintLayout>
