+++ app/src/main/res/layout/list_notice.xml
@@ -5,8 +5,8 @@
 
     <data>
         <variable
-            name="allNoticeList"
-            type="com.example.dmsport_android.feature.notice.model.AllNoticeList" />
+            name="model"
+            type="com.example.dmsport_android.feature.notice.model.Notice" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -30,7 +30,7 @@
                 <TextView
                     android:id="@+id/tv_notice_title1"
                     style="@style/Theme.DMSportAndroid.TextView_notice_title"
-                    android:text="@{allNoticeList.title}"
+                    android:text="@{model.title}"
                     android:layout_marginStart="16dp"
                     android:layout_marginEnd="32dp"
                     app:layout_constraintStart_toStartOf="parent"
@@ -41,7 +41,7 @@
                     android:id="@+id/tv_notice_content1"
                     app:layout_constraintEnd_toEndOf="parent"
                     style="@style/Theme.DMSportAndroid.TextView_notice_content"
-                    android:text="@{allNoticeList.content_preview}"
+                    android:text="@{model.content_preview}"
                     android:layout_marginHorizontal="16dp"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@+id/tv_notice_title1" />
